requirements.txt
fastapi
uvicorn[standard]
tensorflow
numpy
opencv-python
pillow
streamlit
requests
